Type Checking Mode: basic

+++ benchmark/service/task_executor/task_execute_result.py
@@ -66,7 +66,7 @@ class SingleTaskExecuteResult:
     cpu_avg_percent: float
     cpu_samples_count: int
     cpu_sampling_interval: float
-    peak_memory_bytes: float
+    peak_memory_bytes: int
     execution_time: float
     monitor_record_execution_time: float
     output_rows: int
